LayoutTests/transitions/clip-path-path-transitions-expected.txt

CONSOLE MESSAGE: A transition for property clip-path could not be found
CONSOLE MESSAGE: Failed to pause '-webkit-clip-path' transition on element 'path2'
CONSOLE MESSAGE: A transition for property clip-path could not be found
CONSOLE MESSAGE: Failed to pause '-webkit-clip-path' transition on element 'path3'

PASS - "-webkit-clip-path" property for "path1" element at 1s saw something close to: path("M 80 40 l 20 0 l 0 60 l 20 0 l 0 -20 l -50 0 l 0 -20 l 80 0 l 0 60 l -60 0 l 0 -80 Z")
PASS - "-webkit-clip-path" property for "path2" element at 1s saw something close to: path("M 100 40 l 20 0 l 0 60 l 0 -20 l -60 0 l 0 -20 l 80 0 l 0 60 l -60 0 l 0 -80 Z")
PASS - "-webkit-clip-path" property for "path3" element at 1s saw something close to: path(evenodd, "M 100 40 l 20 0 l 0 60 l 20 0 l 0 -20 l -60 0 l 0 -20 l 80 0 l 0 60 l -60 0 l 0 -80 Z")
